Revert "[Refractor] Update admin features."
This reverts commit 8689d57bc46d2c9faa486a4fab8970fbbaab90cb.

+++ Manga/app/src/main/java/nhdphuong/com/manga/features/admin/AdminContract.kt
@@ -8,7 +8,7 @@ interface AdminContract {
         fun updateProgress()
         fun showNumberOfPages(numOfPages: Long)
         fun updateDownloadingStatistics(
-                downloadedPages: Long = 0,
+                downloadedPages: Int = 0,
                 artists: Int = 0,
                 characters: Int = 0,
                 categories: Int = 0,
