+++ Manga/app/src/main/java/nhdphuong/com/manga/features/admin/AdminFragment.kt
@@ -87,7 +87,7 @@ class AdminFragment : Fragment(), AdminContract.View {
     }
 
     override fun updateDownloadingStatistics(
-            downloadedPages: Long,
+            downloadedPages: Int,
             artists: Int,
             characters: Int,
             categories: Int,
